Rewrote Shader->Mesh referencing. Added Player and Scene classes to prepare for modular scene rendering as well as independant player instances in scenes.

+++ LabEngine/Model.h
@@ -40,7 +40,7 @@ class Model
 
 	private:
 		QString name;
-		GLuint shaderID;
+		size_t shaderID;
 		ShaderManager* shaderManager;
 		//TODO: Maybe change to smartpointer!
 		std::vector<Mesh*> meshes;
